talentboost/talentboost/requirements.txt
Django==4.2.9
psycopg2-binary==2.9.9
python-decouple==3.8
django-admin-interface
django-import-export
django-extensions
crispy-bootstrap5
django-compressor
django-filter
django-tinymce
django-registration-redux
django-tables2
django-versatileimagefield
razorpay
scipy
rosetta
simplejson
nltk
pandas
